pyproject.toml
[project]
name = "agent"
version = "0.0.1"
description = "Agent service - RAG, tool execution, background tasks"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt==4.3.0",
    "cryptography>=46.0.3",
    "email-validator>=2.3.0",
    "fastapi>=0.122.0",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "langchain>=1.1.0",
    "langchain-openai>=1.1.2",
    "langgraph>=1.0.4",
    "passlib[bcrypt]==1.7.4",
    "pydantic>=2.12.5",
    "pydantic-core>=2.41.5",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "pyyaml>=6.0.3",
    "psycopg2-binary>=2.9.0",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.44",
    "starlette>=0.50.0",
    "uvicorn>=0.38.0",
    "types-requests>=2.32.4.20260107",
]

[dependency-groups]
dev = [
    "black",
    "isort",
    "flake8",
    "mypy",
    "pre-commit",
]
